move uploadFile methods from howTo store  to common store - use COLLECTION_NAME const instead of hardcoded
